--- tests/test_more_coverage.py
@@ -1,206 +1,0 @@
-"""More tests to reach >90% coverage."""
-
-from datetime import UTC, datetime
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
-from nio import AsyncClient, RoomRedactResponse, RoomSendResponse
-
-from matty import (
-    Message,
-    OutputFormat,
-    _build_message_content,
-    _execute_messages_command,
-    _get_relation,
-    _get_thread_messages,
-    _is_relation_type,
-    _send_message,
-    _send_reaction,
-)
-
-
-class TestReactionsAndRedactions:
-    """Test reaction and redaction functionality."""
-
-    @pytest.mark.asyncio
-    async def test_send_reaction_success(self):
-        """Test successful reaction send."""
-        client = MagicMock(spec=AsyncClient)
-        response = RoomSendResponse(event_id="$reaction123", room_id="!room:matrix.org")
-        client.room_send = AsyncMock(return_value=response)
-
-        result = await _send_reaction(client, "!room:matrix.org", "$msg123", "👍")
-        assert result is True
-
-        # Verify reaction content
-        call_args = client.room_send.call_args
-        content = call_args[1]["content"]
-        assert content["m.relates_to"]["rel_type"] == "m.annotation"
-        assert content["m.relates_to"]["event_id"] == "$msg123"
-        assert content["m.relates_to"]["key"] == "👍"
-
-    @pytest.mark.asyncio
-    async def test_send_reaction_failure(self):
-        """Test reaction send failure."""
-        client = MagicMock(spec=AsyncClient)
-        client.room_send = AsyncMock(side_effect=Exception("Network error"))
-
-        result = await _send_reaction(client, "!room:matrix.org", "$msg123", "👍")
-        assert result is False
-
-    @pytest.mark.asyncio
-    async def test_room_redact_success(self):
-        """Test successful redaction via room_redact."""
-        client = MagicMock(spec=AsyncClient)
-        response = RoomRedactResponse(event_id="$redaction123", room_id="!room:matrix.org")
-        client.room_redact = AsyncMock(return_value=response)
-
-        # Directly test the client method since _send_redaction doesn't exist
-        result = await client.room_redact("!room:matrix.org", "$msg456", reason="Mistake")
-        assert result.event_id == "$redaction123"
-
-        # Verify redaction was called
-        client.room_redact.assert_called_once_with("!room:matrix.org", "$msg456", reason="Mistake")
-
-
-class TestThreadMessages:
-    """Test thread message handling."""
-
-    @pytest.mark.asyncio
-    async def test_get_thread_messages_simple(self):
-        """Test getting thread messages."""
-        from nio import RoomMessageText
-
-        client = MagicMock(spec=AsyncClient)
-
-        # Create mock thread messages
-        msg1 = MagicMock(spec=RoomMessageText)
-        msg1.sender = "@alice:matrix.org"
-        msg1.body = "Thread root"
-        msg1.server_timestamp = 1704110400000
-        msg1.event_id = "$root123"
-        msg1.source = {
-            "content": {
-                "body": "Thread root",
-                "msgtype": "m.text",
-                "m.relates_to": {"rel_type": "m.thread", "event_id": "$root123"},
-            }
-        }
-
-        mock_response = MagicMock()
-        mock_response.chunk = [msg1]
-
-        client.room_messages = AsyncMock(return_value=mock_response)
-
-        with patch("matty._get_or_create_handle", return_value="m1"):
-            messages = await _get_thread_messages(client, "!room:matrix.org", "$root123")
-
-        assert len(messages) > 0
-
-
-class TestMessageContent:
-    """Test message content building with various options."""
-
-    def test_build_message_with_all_options(self):
-        """Test building message with all relations."""
-        content = _build_message_content(
-            "Test message",
-            formatted_body="<b>Test message</b>",
-            mentioned_user_ids=["@user:matrix.org"],
-            thread_root_id="$thread123",
-            reply_to_id="$reply456",
-        )
-
-        assert content["body"] == "Test message"
-        assert content["formatted_body"] == "<b>Test message</b>"
-        assert content["format"] == "org.matrix.custom.html"
-        assert content["m.mentions"]["user_ids"] == ["@user:matrix.org"]
-        assert content["m.relates_to"]["rel_type"] == "m.thread"
-        assert content["m.relates_to"]["event_id"] == "$thread123"
-        assert content["m.relates_to"]["m.in_reply_to"]["event_id"] == "$reply456"
-
-
-class TestThreadExecution:
-    """Test thread-related command execution."""
-
-    @pytest.mark.asyncio
-    async def test_execute_messages_with_thread_success(self, capsys):
-        """Test messages command with thread parameter."""
-        with patch("matty._create_client") as mock_create:
-            client = MagicMock(spec=AsyncClient)
-            mock_create.return_value = client
-
-            with (
-                patch("matty._login", return_value=True),
-                patch("matty._sync_client"),
-                patch("matty._find_room", return_value=("!room:matrix.org", "Test Room")),
-                patch("matty._resolve_thread_id", return_value=("$thread123", None)),
-            ):
-                messages = [
-                    Message(
-                        sender="@alice:matrix.org",
-                        content="Thread message",
-                        timestamp=datetime.now(UTC),
-                        room_id="!room:matrix.org",
-                        event_id="$msg1",
-                        handle="m1",
-                    )
-                ]
-                with patch("matty._get_thread_messages", return_value=messages):
-                    client.close = AsyncMock()
-
-                    # _execute_messages_command doesn't have thread parameter
-                    await _execute_messages_command(
-                        "Test Room",
-                        10,
-                        "user",
-                        "pass",
-                        OutputFormat.simple,
-                    )
-
-        captured = capsys.readouterr()
-        assert "Test Room" in captured.out or "Thread" in captured.out
-
-
-class TestEdgeCases:
-    """Test additional edge cases."""
-
-    def test_get_relation_none(self):
-        """Test getting relation when not present."""
-        content = {"body": "message"}
-        assert _get_relation(content) is None
-
-    def test_get_relation_present(self):
-        """Test getting relation when present."""
-        content = {"m.relates_to": {"rel_type": "m.thread"}}
-        relation = _get_relation(content)
-        assert relation["rel_type"] == "m.thread"
-
-    def test_is_relation_type_true(self):
-        """Test checking relation type matches."""
-        content = {"m.relates_to": {"rel_type": "m.thread"}}
-        assert _is_relation_type(content, "m.thread") is True
-        assert _is_relation_type(content, "m.annotation") is False
-
-    def test_is_relation_type_no_relation(self):
-        """Test checking relation type when no relation."""
-        content = {"body": "message"}
-        assert _is_relation_type(content, "m.thread") is False
-
-    @pytest.mark.asyncio
-    async def test_send_message_with_formatted_body(self):
-        """Test sending message with formatted body."""
-        client = MagicMock(spec=AsyncClient)
-        # Mock the rooms attribute
-        client.rooms = {}
-        response = RoomSendResponse(event_id="$formatted123", room_id="!room:matrix.org")
-        client.room_send = AsyncMock(return_value=response)
-
-        # _send_message doesn't have formatted_body parameter, it parses HTML from message
-        result = await _send_message(client, "!room:matrix.org", "Plain text")
-        assert result is True
-
-        # Verify basic message was sent
-        call_args = client.room_send.call_args
-        content = call_args[1]["content"]
-        assert content["body"] == "Plain text"
